python-script-flaskAPI/requirements.txt
astroid==2.4.2
click==7.1.2
colorama==0.4.4
cycler==0.10.0
Flask==1.1.2
gunicorn==20.0.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
numpy==1.19.5
Pillow==8.1.0
pybase64==1.1.4
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
Flask-Cors==3.0.10